chore: move compilation.rs into mod.rs
